rename tile16.py to _tile16.py and remove from public API
The module is internal — all its symbols are already underscore-prefixed.
Remove from __all__ and __getattr__ in simt/__init__.py so it's not
accessible via qd.simt.tile16.
