+++ python/quadrants/lang/simt/__init__.py
@@ -2,12 +2,4 @@
 
 from quadrants.lang.simt import block, grid, subgroup, warp
 
-__all__ = ["warp", "subgroup", "block", "grid", "tile16"]
-
-
-def __getattr__(name):
-    if name == "tile16":
-        from quadrants.lang.simt import tile16  # noqa: I001  # pylint: disable=import-outside-toplevel
-
-        return tile16
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+__all__ = ["warp", "subgroup", "block", "grid"]
